Move video functions to their own blueprint [REFACTORING]
