+++ myclimbz/templates/form_sort_videos.html
@@ -10,13 +10,13 @@
         {% for video in videos %}
         <div class="list-group-item list-group-item-action d-flex justify-content-center" data-fname="{{ video }}">
             <video height="320" preload="metadata" class="me-2">
-                <source src="{{ url_for('climbs.serve_file', filetype='videos', filename=video) }}" alt="{{ video }}">
+                <source src="{{ url_for('videos.serve_file', filetype='videos', filename=video) }}" alt="{{ video }}">
             </video>
         </div>
         {% endfor %}
     </div>
 
-    <form id="sort-form" method="POST" action="{{ url_for('climbs.sort_videos') }}">
+    <form id="sort-form" method="POST" action="{{ url_for('videos.sort_videos') }}">
         {{ form.hidden_tag() }}
         <div style="display:none;"> {{ form.videos_sorted() }} </div>
         <input type="submit" class="btn btn-success" name="btn" value="Continue">
